Finished PL on console

+++ levels/pl/pl_mlift/geo.inc.c
@@ -4,7 +4,7 @@ const GeoLayout pl_mlift_geo[] = {
 	GEO_NODE_START(),
 	GEO_OPEN_NODE(),
 		GEO_DISPLAY_LIST(LAYER_OPAQUE, pl_mlift_pl_mlift_mesh_layer_1),
-		GEO_TRANSLATE_NODE_WITH_DL(LAYER_OPAQUE, 0, 30, 0, pl_mlift_001_object_00C3BF9C_mesh_layer_1),
+		GEO_TRANSLATE_NODE_WITH_DL(LAYER_OPAQUE, (s16)0.0, (s16)30.0, (s16)0.0, pl_mlift_001_object_00C3BF9C_mesh_layer_1),
 	GEO_CLOSE_NODE(),
 	GEO_END(),
 };
